% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launchMACERShiny}
\alias{launchMACERShiny}
\title{Launch MACERShiny}
\usage{
launchMACERShiny()
}
\value{
There are no values or files returned from this function
}
\description{
This function launches the MACERShiny Application
}
\details{
This function launches a MACERShiny Application which allows a user to run the DBTC
functions and process high throughput sequencing data.
}
\note{
This is a wrapper function which launches the MACERShiny package as a shiny
application in the systems default browser program
}
\examples{
\dontrun{
launchMACERShiny()
}

}
\references{
<https://github.com/rgyoung6/MACER>
Young, R. G., Hanner, R. H. (Submitted October 2023). Title Here. Biodiversity Data Journal.
}
\author{
Robert G. Young
}
